requirements.txt
Flask==3.0.2
gradio-client==1.8.0
Pillow==10.2.0
requests==2.31.0
gunicorn
